requirements.txt
opencv-python>=4.5.5.64
numpy>=1.22.3
pillow>=7.0.0
bs4>=0.0.1
beautifulsoup4>=4.10.0
tqdm>=4.62.3
imgaug>=0.4.0
colorama>=0.4.4
torch>=1.10.1+cu113
GoogleDriveDownloader
